[C] Add ReferenceBlock positions
For JIRA: "EPO-5803 #IN-REVIEW #comment Add ReferenceBlock and position"
