+++ src/data/pages/orbital-interpretation-1.json
@@ -20,6 +20,9 @@
       "button": {
         "text": "Planetary Orbital Sizes"
       },
+      "options": {
+        "position": "right"
+      },
       "tables": [
         {
           "id": "1",
