+++ src/data/pages/semi-major-axis.json
@@ -23,6 +23,13 @@
   ],
   "reference": [
     {
+      "layout": {
+        "col": "left",
+        "row": "bottom"
+      },
+      "options": {
+        "position": "left"
+      },
       "title": "Planetary Orbital Sizes",
       "button": {
         "text": "Planetary Orbital Sizes"
